requirements.txt
flask>=3.1.2
insightface>=0.7.3
onnxruntime>=1.23.0
opencv-python-headless>=4.12.0.88
paho-mqtt>=1.6.1
watchdog>=6.0.0
